+++ tests/Frameworks/Lumen/Version_5_2/resources/views/simple_view.blade.php
@@ -1,0 +1,7 @@
+<!doctype html>
+<html>
+    <head></head>
+    <body>
+        <h1>Simple view</h1>
+    </body>
+</html>