+++ tests/Frameworks/Lumen/Version_5_6/app/Http/Controllers/ExampleController.php
@@ -14,5 +14,18 @@ class ExampleController extends Controller
         //
     }
 
-    //
+    public function simple()
+    {
+        return 'simple';
+    }
+
+    public function simpleView()
+    {
+        return view('simple_view');
+    }
+
+    public function error()
+    {
+        throw new \Exception('Controller error');
+    }
 }