+++ src/DDTrace/Tracer.php
@@ -25,7 +25,7 @@ final class Tracer implements TracerInterface
 {
     use LoggingTrait;
 
-    const VERSION = '0.23.0';
+    const VERSION = '0.24.0';
 
     /**
      * @var Span[][]